+++ petanque.h
@@ -9,7 +9,7 @@ class Petanque : public Defense
 {
 public:
     Petanque();
-    Petanque(const int niveau,const int x,const int y,QGraphicsItem *parent=0);
+    Petanque(const double niveau,const int x,const int y,QGraphicsItem *parent=0);
 
     double attaquer();
     void ameliorer();
